setup()

+++ c/cli.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <stddef.h>
+
 typedef enum io_type_e {
   NONE,
   RAND_READ,
@@ -12,8 +14,8 @@ typedef struct config_s {
   const char *program;
   const char *workdir;
   io_type_e io_type;
-  unsigned long long bs;
-  unsigned long long count;
+  size_t bs;
+  size_t count;
 } config_s;
 
 config_s cli_parse(int argc, char *argv[]);
